PDF Article
In order to achieve stump healing after trans-tibial amputation, various methods are applied, such as soft dressings followed by elastic wrapping of the stump, rigid dressings, semi-rigid dressings, and more recently the application of silicon or gel-liners. A systematic literature search was performed to identify the optimal post-amputation management. The methodological quality of the studies was systematically evaluated by using a predefined list of criteria. Only 11 controlled studies were identified and evaluated for their methodological quality. From these studies, no studies were classified as A-level studies, whereas three were classified as B-level, and 8 were classified as C-level studies. Relevant literature appears heterogeneous with respect to patient selection, intervention and outcome measures. Despite the large variability of included studies, this review reveals a trend in favour of rigid and semi-rigid dressings for achieving stump healing and reduction of stump volume. No conclusions can be drawn with regard to the effect on functional outcome. The literature is not conclusive on the effects of early weight bearing on stump healing, volume reduction, and functional outcome. More research is needed for the development of evidence-based clinical practice guidelines concerning management after trans-tibial amputation.